feat: rrule scheduling
This works but is rough around the edges. The logic calculates the next
occurence of a task in order to determine when to sleep. I believe that
this could be a potentially problematic implementation because it could
require a heavy amount of iteration as the dtstart of the rule becomes
further in the past. A future refactor will likely be required to more
efficiently perform this by avoiding extra iteration. My current
intuition says that the iteration could be performed at launch time and
the scheduler can keep a handle to the iteration in order to only
calculate the next value each time, rather than iterate from the start
each time.